Duplicating Digglet's Cave warps
Digglet's Cave can be approached from either Route 2 or Route 11, and thus the "Previous Map" could be set to either one as the player travels through the cave. The use of only one side in the map meant that sometimes Red would fail to warp to FireRed if coming from the Route 2 side. This has been rectified by duplicating the mapping for each side.

The mapping back from FireRed is unaltered, though it technically has a 50% chance to assign an incorrect value back to Red on warping back. It'll be up to TPP to dig out when the assumption is wrong!